Error Handling Fix#1

+++ controllers/users/adminController.js
@@ -8,6 +8,41 @@ module.exports.signUp = async (req, res) => {
     const phoneNo = req.body.phoneNo
     const postName = req.body.postName
     const superAdminId = req.userData.uid
+    
+    const requiredFields = ["name", "email", "phoneNo", "postName"]
+
+    // check if all the required fields are present by looping
+    for(let i=0; i<requiredFields.length; i++){
+        if(!req.body.hasOwnProperty(requiredFields[i])){
+            return res.status(400).json({
+                message: `${requiredFields[i]} is required`
+            })
+        } else if(typeof req.body[requiredFields[i]] !== "string"){
+            return res.status(400).json({
+                message: `${requiredFields[i]} should be a string`
+            })
+        } else if(req.body[requiredFields[i]].trim().length === 0){
+            return res.status(400).json({
+                message: `${requiredFields[i]} can't be empty`
+            })
+        }
+    }
+
+    // check if the email is valid
+    const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+    if(!emailRegex.test(email)) {
+        return res.status(400).json({
+            message: 'Email is not valid'
+        })
+    }
+
+    // check if the phone number is valid
+    const phoneRegex = /^[0-9]{10}$/;
+    if(!phoneRegex.test(phoneNo)) {
+        return res.status(400).json({
+            message: 'Phone number is not valid'
+        })
+    }
 
     firestore.collection('users').doc(superAdminId).get()
     .then(user => {
@@ -87,7 +122,7 @@ module.exports.getAdmin = (req,res) => {
         });
     })
     .catch((error) => {
-        console.error('Error getting admin:', error);
+        console.error(error);
         return res.status(500).json({
             message: 'Internal Server Error',
         });
@@ -98,14 +133,58 @@ module.exports.updateAdmin = (req,res) => {
     const adminuid = req.userData.uid
     const updatedData = req.body.updateData
 
-    firestore.collection('users').doc(adminuid).update(updatedData)
-    .then(() => {
+    const prohibitedFields = ["accessLevel", "mailID", "roleName", "isSuspended", "plantID", "dateAdded"]
+    const updateableFields = ["name", "postName", "phoneNo"]
+
+    // check if the updatedData object contains any prohibited fields
+    for(let i=0; i<prohibitedFields.length; i++){
+        if(updatedData.hasOwnProperty(prohibitedFields[i])){
+            return res.status(400).json({
+                message: `${prohibitedFields[i]} can't be updated`
+            })
+        }
+    }
+
+    // check if the updatedData object contains any updateable fields and if they are valid strings and non empty
+    for(let i=0; i<updateableFields.length; i++){
+        if(updatedData.hasOwnProperty(updateableFields[i])){
+            if(typeof updatedData[updateableFields[i]] !== "string"){
+                return res.status(400).json({
+                    message: `${updateableFields[i]} should be a string`
+                })
+            } else if(updatedData[updateableFields[i]].trim().length === 0){
+                return res.status(400).json({
+                    message: `${updateableFields[i]} can't be empty`
+                })
+            }
+        }
+    }
+
+    // check if the updatedData contains any fields which are not in updateableFields and in prohibitedFields
+    for(let key in updatedData){
+        if(!updateableFields.includes(key) && !prohibitedFields.includes(key)){
+            return res.status(400).json({
+                message: `${key} is not a valid field to update`
+            })
+        }
+    }
+
+    firestore.collection('users').doc(adminuid).get()
+    .then(async admin => {
+        if(!admin.exists){
+            return res.status(404).json({
+                message: "Admin not found"
+            })
+        }
+
+        await firestore.collection('users').doc(adminuid).update(updatedData)
+
         return res.status(200).json({
             message: 'Admin updated successfully',
         });
     })
     .catch((error) => {
-        console.error('Error updating user:', error);
+        console.error(error);
         return res.status(500).json({
             message: 'Internal Server Error',
         });
@@ -120,6 +199,13 @@ module.exports.deleteAdmin = (req,res) => {
     .then(async user => {
         if(user.exists && user.get('roleName') === "superAdmin"){
 
+            // find the admin to be deleted
+            const admin = await firestore.collection('users').doc(adminuid).get()
+            if(!admin.exists){
+                return res.status(404).json({
+                    message: "Admin not found"
+                })
+            }
             await firebase.auth().deleteUser(adminuid)
             
             await firestore.collection('users').doc(adminuid).delete()
@@ -130,7 +216,7 @@ module.exports.deleteAdmin = (req,res) => {
 
         } else {
             return res.status(401).json({
-                message: "Only an admin can delete users"
+                message: "Only a super admin can delete users"
             })
         }
     })
